+++ sdk/storage/Azure.Storage.DataMovement/tests/Shared/MemoryStorageResourceItem.cs
@@ -21,6 +21,8 @@ namespace Azure.Storage.DataMovement.Tests
 
         protected internal override DataTransferOrder TransferType => DataTransferOrder.Unordered;
 
+        protected internal override long MaxSupportedSingleTransferSize => long.MaxValue;
+
         protected internal override long MaxSupportedChunkSize => long.MaxValue;
 
         protected internal override long? Length => Buffer.Length;