+++ sdk/cosmosdb/Azure.ResourceManager.CosmosDB/CHANGELOG.md
@@ -1,16 +1,22 @@
 # Release History
 
-## 1.4.0-beta.11 (Unreleased)
+## 1.4.0-beta.12 (Unreleased)
 
 ### Features Added
-
-- Exposed `JsonModelWriteCore` for model serialization procedure.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.4.0-beta.11 (2024-11-01)
+
+### Features Added
+
+- Upgraded api-version tag from 'package-preview-2024-05' to 'package-preview-2024-09'. Tag detail available at https://github.com/Azure/azure-rest-api-specs/blob/4d065866a422257746306b352cb34fd5c98d5754/specification/cosmos-db/resource-manager/readme.md
+- Exposed `JsonModelWriteCore` for model serialization procedure.
+
 
 ## 1.4.0-beta.10 (2024-06-25)